# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Svet API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/user/create": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "description": "Регистрация",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почта"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль"
                    },
                    {
                        "name": "login",
                        "in": "query",
                        "description": "Логин"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Телефон"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/verified": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "description": "Подтверждение регистрации по SMS (на данный момент по почте)",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID пользователя"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Код подтверждения"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/authorize": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "description": "Авторизация",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-Mail"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/recovery/password": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "description": "Запрос (отправка кода) на смену пароля",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-Mail для восстановления"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/reset/password": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "description": "Подтверждение смены пароля",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-Mail"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Код из E-Mail сообщения"
                    },
                    {
                        "name": "password1",
                        "in": "query",
                        "description": "Пароль 1"
                    },
                    {
                        "name": "password2",
                        "in": "query",
                        "description": "Пароль 2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        }
    }
}